+++ src/org/hyperic/hq/appdef/shared/PlatformValue.java
@@ -675,6 +675,6 @@ public class PlatformValue extends AppdefResourceValue
     }
 
     public AppdefEntityID getEntityId() {
-        return AppdefEntityID.newPlatformID(getId().intValue());       
+        return AppdefEntityID.newPlatformID(getId());       
     }
 }
